bus: organize modules
